패키지제이슨 설치 스토어 연동 db.json 데이터 추가 메인리스트모듈 생성(적용x)
