+++ 20questions/db.json
@@ -1,0 +1,44 @@
+{
+  "data": [
+    {
+      "nickname": "수민",
+      "quizId": 1,
+      "title": "뭘까요?",
+      "count": 7,
+      "category": 1,
+      "date": "2020 / 08 / 13"
+    },
+    {
+      "nickname": "수민",
+      "quizId": 2,
+      "title": "뭘까요?",
+      "count": 7,
+      "category": 1,
+      "date": "2020 / 08 / 13"
+    },
+    {
+      "nickname": "종수",
+      "quizId": 3,
+      "title": "뭘까요?",
+      "count": 7,
+      "category": 1,
+      "date": "2020 / 08 / 13"
+    },
+    {
+      "nickname": "수민",
+      "quizId": 4,
+      "title": "뭘까요?",
+      "count": 7,
+      "category": 1,
+      "date": "2020 / 08 / 13"
+    },
+    {
+      "nickname": "종민",
+      "quizId": 5,
+      "title": "뭘까요?",
+      "count": 7,
+      "category": 1,
+      "date": "2020 / 08 / 13"
+    }
+  ]
+}
