+++ 20questions/src/index.js
@@ -4,16 +4,16 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
 import { BrowserRouter } from "react-router-dom";
-// import { Provider } from "react-redux";
-// import store from "./redux/storeConfig";
+import { Provider } from "react-redux";
+import store from "./redux/config/storeConfig";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <div>
     <BrowserRouter>
-      {/* <Provider store={store}> */}
+      <Provider store={store}>
       <App />
-      {/* </Provider> */}
+      </Provider>
     </BrowserRouter>
   </div>
 );
